dedicated source file for common events tests
